requirements.txt
django
django-rest-framework
requests
django-cors-headers
faker
web3
python-dotenv
